Move initAndCreateThreePlayerGame into AITest
